Drop session.cookies and CookiesProxy, add session.cookie_jar
