+++ tests/test_test_utils.py
@@ -159,9 +159,9 @@ def test_client_websocket(loop, test_client):
 
 @asyncio.coroutine
 def test_client_cookie(loop, test_client):
-    assert not test_client.session.cookies
+    assert not test_client.session.cookie_jar
     yield from test_client.get("/cookie")
-    cookies = list(test_client.session.cookies)
+    cookies = list(test_client.session.cookie_jar)
     assert cookies[0].key == 'cookie'
     assert cookies[0].value == 'val'
